+++ lib/uglifier.rb
@@ -83,7 +83,8 @@ class Uglifier
       :keep_fargs => false, # Preserve unused function arguments
       :keep_fnames => false, # Do not drop names in function definitions
       :passes => 1, # Number of times to run compress. Raising the number of passes will increase compress time, but can produce slightly smaller code.
-      :keep_infinity => false # Prevent compression of Infinity to 1/0
+      :keep_infinity => false, # Prevent compression of Infinity to 1/0
+      :side_effects => true, # Pass false to disable potentially dropping functions marked as "pure" using pure comment annotation. See UglifyJS documentation for details.
     }, # Apply transformations to code, set to false to skip
     :parse => {
       :bare_returns => false, # Allow top-level return statements.