+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - add shebang output option
 - add keep_infinity compress option
 - add quote_style output option
+- add side_effects compress option
 
 Dropped options: